Had to switch from QR scanning library due to the model being used to read the qr codes; apparently the one used on shodan.io was too complex for the previous one. Now also auto-login. Cleaned up all em unnecessary libraries.
