+++ pubspec.yaml
@@ -16,13 +16,11 @@ dependencies:
   path: ^1.8.3
   path_provider: ^2.1.1
   provider: ^6.1.1
-  rename_app: ^1.3.1
-  shared_preferences: ^2.2.2
   window_manager: ^0.3.7
   url_launcher: 
   step_progress_indicator: ^1.0.2
-  mobile_scanner: ^3.5.5
   flutter_secure_storage: ^9.0.0
+  qr_mobile_vision: ^4.1.4
 
 dev_dependencies:
   flutter_test:
